# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IPN-DEV — Sistema de Inventario Daluzed Pastelería",
        "description": "API REST para gestión de inventario, producción y logística de Daluzed Pastelería. Proyecto Nuclear 3 · Corporación Universitaria Alexander von Humboldt (CUE).",
        "contact": {
            "name": "Equipo IPN-DEV",
            "email": "admin@inventario.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/bodegas": {
            "get": {
                "tags": [
                    "Catálogo - Bodegas"
                ],
                "summary": "Listar bodegas",
                "operationId": "dff03f0248a8bb557be8a6609423ffd1",
                "responses": {
                    "200": {
                        "description": "Listado de bodegas"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso (requiere bodegas.leer)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Catálogo - Bodegas"
                ],
                "summary": "Crear bodega",
                "operationId": "6aa023d6667e47b0dfcb41b78d2ac5f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "tipo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Bodega Norte"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "enum": [
                                            "principal",
                                            "produccion",
                                            "ventas"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bodega creada"
                    },
                    "403": {
                        "description": "Sin permiso (requiere bodegas.escribir)"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/bodegas/{id}": {
            "get": {
                "tags": [
                    "Catálogo - Bodegas"
                ],
                "summary": "Ver bodega",
                "operationId": "aadd46c5d9f025797ebfb2805ba9a2ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la bodega"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Catálogo - Bodegas"
                ],
                "summary": "Actualizar bodega",
                "operationId": "686052f62a6e1b0aa024d8f88f1de99b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "activa": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bodega actualizada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Bearer generado por Laravel Sanctum al hacer login. Incluirlo en el header: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}